# server/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Mask-shop API",
    "description": "Mask-shop API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "mask-shop.kro.kr/v1",
  "basePath": "/api",
...
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/category": {
      "post": {
        "tags": [
          "category"
        ],
        "description": "Create new category in system",
        "parameters": [
          {
            "name": "category name",
            "in": "body",
            "description": "User that we want to create",
            "schema": {
              "$ref": "#/definitions/category"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New user is created",
            "schema": {
              "$ref": "#/definitions/category"
            }
          }
        }
      },
      "get": {
        "tags": [
          "category"
        ],
        "summary": "Get all users in system",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/category"
            }
          }
        }
      }
    },
...
    "/subcategory": {
      "post": {
        "tags": [
          "subcategory"
        ],
        "description": "Create new subcategory in system",
        "parameters": [
          {
            "name": "subcategory name",
            "in": "body",
            "description": "subcategory that we want to create",
            "schema": {
              "$ref": "#/definitions/subcategory"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New subcategory is created",
            "schema": {
              "$ref": "#/definitions/subcategory"
            }
          }
        }
      },
      "get": {
        "tags": [
          "subcategory"
        ],
        "summary": "Get all subcategory in system",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/subcategory"
            }
          }
        }
      }
    },
...
    },"/order": {
        "post": {
          "tags": [
            "order"
          ],
          "description": "Create new order in system",
          "parameters": [
            {
              "name": "order",
              "in": "body",
              "description": "User that we want to create",
              "schema": {
                "$ref": "#/definitions/order"
              }
            }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "New order is created",
              "schema": {
                "$ref": "#/definitions/order"
              }
            }
          }
        }},
      "/order/list/:id": {
        "parameters": [
          {
            "name": "�ֹ���ȣ ",
            "in": "path",
            "required": true,
            "description": "ID of order that we want to find",
            "type": "string"
          }
        ],
        "get": {
          "tags": [
            "order"
          ],
          "summary": "Get order with given ID",
          "responses": {
            "200": {
              "description": "order is found",
              "schema": {
                "$ref": "#/definitions/order"
              }
            }
          }
        },
        "delete": {
          "summary": "Delete order with given ID",
          "tags": [
            "order"
          ],
          "responses": {
            "200": {
              "description": "order is deleted",
              "schema": {
                "$ref": "#/definitions/order"
              }
            }
          }
        }
      },
        "/order/:id": {
        "parameters": [
          {
            "name": "ȸ�����̵� ",
            "in": "path",
            "required": true,
            "description": "ID of order that we want to find",
            "type": "string"
          }
        ],
        "get": {
          "tags": [
            "order"
          ],
          "summary": "Get order with given ID",
          "responses": {
            "200": {
              "description": "order is found",
              "schema": {
                "$ref": "#/definitions/order"
              }
            }
          }
        },
        "delete": {
          "summary": "Delete order with given ID",
          "tags": [
            "order"
          ],
          "responses": {
            "200": {
              "description": "order is deleted",
              "schema": {
                "$ref": "#/definitions/order"
              }
            }
          }
        }
      },
       "/auth/login": {
        "post": {
          "tags": [
            "auth"
          ],
          "description": "Create new token in system",
          "parameters": [
            {
              "name": "order",
              "in": "body",
              "description": "token that we want to create",
              "schema": {
                "$ref": "#/definitions/login"
              }
            }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "New token is created",
              "schema": {
                "$ref": "#/definitions/token"
              }
            }
          }
        }
      },
      "/auth/me": {
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of order that we want to find",
            "type": "string"
          }
        ],
        "get": {
          "tags": [
            "auth"
          ],
          "summary": "Get user with given token",
          "responses": {
            "200": {
              "description": "user is found",
              "schema": {
                "$ref": "#/definitions/user"
              }
            }
          }
        }
      },
      "/auth/refresh": {
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "description": "make refresh token",
            "type": "string"
          }
        ],
        "get": {
          "tags": [
            "auth"
          ],
          "summary": "make refresh token",
          "responses": {
            "200": {
              "description": "refresh token",
              "schema": {
                "$ref": "#/definitions/token"
              }
            }
          }
        }
      },
...
  },
  "definitions": {
    "category": {
      "required": [
        "category"
      ],
      "properties": {
        "category": {
          "type": "string",
          "uniqueItems": true
        },
        "subcategory": {
          "type": "[string]"
        }
      }
    },
...
    "subcategory": {
      "required": [
        "category"
      ],
      "properties": {
        "category": {
          "type": "string",
          "uniqueItems": true
        },
        "subcategory": {
          "type": "[string]"
        }
      }
    },
...
    "token": {
        "required": [
          "success",
          "message",
          "errors",
          "data"
        ],
        "properties": {
            "success": {
                "type": "string",
                "uniqueItems": true
            },
            "message": {
                "type": "string"
            },
            "errors": {
                "type": "string"
            },
            "data": {
                "type": "string"
            }
        }
      },
      "login": {
        "required": [
          "accountid",
          "password"
        ],
        "properties": {
            "accountid": {
                "type": "string",
                "uniqueItems": true
            },
            "password": {
                "type": "string"
            }
        }
      },
...
       "order": {
        "required": [
          "orderNum",
          "cycle",
          "price",
          "productName",
          "productCount",
          "accountid",
          "time",
          "invoiceNum",
          "deliver"
        ],
        "properties": {
            "orderNum": {
                "type": "string",
                "uniqueItems": true
            },
            "cycle": {
                "type": "int"
            },
            "price": {
                "type": "int"
            },
            "productName": {
                "type": "string"
            },
            "productCount": {
                "type": "int"
            },
            "accountid": {
                "type": "string"
            },
            "time": {
                "type": "string"
            },
            "invoiceNum": {
                "type": "int"
            },
            "deliver": {
                "type": "string"
            }
        }
      },
...
  }
}